requirements.txt
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-HTMLmin==2.0.0
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
htmlmin==0.1.12
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
lesscpy==0.14.0
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
ply==3.11
pycountry==19.8.18
python-slugify==4.0.0
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
text-unidecode==1.3
Unidecode==1.1.1
urllib3==1.25.9
Werkzeug==1.0.1
xxhash==1.4.3
